Refactor question views to use only a single HTTP endpoint.
This change results in a simplification of the logic by putting all the
code that controls the flow between the different stages in a single place.

The old code verified whether the data we see in the session and the DB is
consistent with the stage we are at according to the URL, and if it wasn't,
it redirected to the right endpoint.  These checks were spread across
multiple functions.

The new version has only a single endpoint.  We load all data first and
delegate to the correct view based on the data.

In addition to being easier to understand and maintain, the new version
also gives the flexibility to support alternative ways of displaying the
rationales.