Return numeric labels as strings.
Labels should be strings.  The old version was inconsistent:
get_choice_label_iter() always returned strings, while get_choice_label()
returned strings for alphabetic labels and ints for numeric labels.  This
resulted in a display bug, because the values returned by the two functions were
compared for equality to decide which answer to highlight as first and second
choice.